Use more correct size_t/ssize_t

+++ src/cmd/ksh93/include/national.h
@@ -33,6 +33,6 @@
 #endif /* SHOPT_MULTIBYTE */
 
 extern ptrdiff_t sh_strchr(const char*,const char*);
-extern unsigned int sh_strwidth(const char*);
+extern size_t sh_strwidth(const char*);
 
 #endif /* _national_h_defined */
